Extract seed setting control to IHasSeed
